backend/requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
django-admin-interface==0.28.6
django-admin-tools==0.9.3
django-appconf==1.0.6
django-colorfield==0.11.0
django-imagekit==5.0.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-redis==5.4.0
django-markdownfield==0.11.0
django-autocomplete-light==3.11.0
django-daterange-filter==1.3.0
redis==5.0.4
msgpack==1.0.8
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
pilkit==3.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-slugify==8.0.4
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn[standard]==0.29.0
requests==2.31.0
celery==5.4.0
celery-singleton==0.3.1
django-celery-beat==2.6.0
channels==4.1.0
channels-redis==4.2.0
PyYAML